refactor(assertions): extract shared script normalization logic
The Python and Ruby assertion handlers had ~120 lines of identical
normalization logic (XOR inverse, threshold comparison, snake_case
mapping, JSON parsing). Extract into scriptResultNormalization.ts
so this logic exists in exactly one place.
